+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### Added
 
+* `pe.GrammarWarning` ([#7])
 * `pe.actions.Action`
 * `pe.actions.Call`
 * `pe.actions.Raw`
 * `pe.actions.Bind`
 * `pe.actions.Getter` (replaces `pe.actions.first` and `pe.actions.last`)
+* `pe.actions.Warn` ([#7])
 * `pe.operators.Debug` ([#9])
 
 ### Removed
@@ -25,6 +27,7 @@
 * `pe.compile()` no longer takes an open file or a single definition as input
 * Grammar vs Expression parsing is more consistent internally
 * Packrat parser handles the Debug operator ([#9])
+* Grammar parser now warns on unlikely range ([#7])
 
 
 ### Fixed
@@ -43,4 +46,5 @@ descent parser and a work-in-progress state-machine parser.
 [unreleased]: ../../tree/develop
 [v0.1.0]: ../../releases/tag/v0.1.0
 
+[#7]: https://github.com/goodmami/pe/issues/7
 [#9]: https://github.com/goodmami/pe/issues/9